je crois que j'ai fix le probleme sur $lol"sdsdsds"$lol"   "$lol $lol"Sdsdsdsd"$lol"   "$lol mais c'est fix assez gros qui peut breaker les autres tests

+++ src/manage_count.c
@@ -19,9 +19,9 @@ void	manage_tab(t_quote *q)
 	q->decr_tab = 0;
 }
 
-void	increment_q_counter_w_tab(size_t *counter, t_quote *q)
+void	increment_q_counter_w_tab(int *counter, t_quote *q)
 {
-	size_t temp_next_quote;
+	int	temp_next_quote;
 	int	z;
 
 	z = q->z;
