+++ src/type_func_problem.c
@@ -12,7 +12,7 @@
 
 #include "minishell.h"
 
-void before_tok(t_quote *q, size_t *i, size_t *quote_counter, char *s)
+void before_tok(t_quote *q, size_t *i, int *quote_counter, char *s)
 {
 	if (!q->count_next_quote)
 	{
@@ -30,7 +30,7 @@ void before_tok(t_quote *q, size_t *i, size_t *quote_counter, char *s)
 	}
 }
 
-void	skip_non_v_quote(char *s, int *temp_value, size_t *quote_counter, size_t *i)
+void	skip_non_v_quote(char *s, int *temp_value, int *quote_counter, size_t *i)
 {
 	(*i)++;
 	while(s[(*i)] && *temp_value)
